requirements.txt
streamlit
serpapi==0.1.4
beautifulsoup4
sumy
requests
python-dotenv
wikipedia
duckduckgo-search
google-search-results
fpdf
python-docx
